graphrag-ui/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.32.1
pandas==2.3.3
pydantic==2.10.4
python-multipart==0.0.20
aiofiles==24.1.0
pyyaml==6.0.2
pyarrow==18.1.0

# GraphRAG 及其核心依賴
datashaper>=0.0.49
environs>=11.0.0

# GraphRAG LLM 依賴
openai>=1.37.1
nltk==3.9.1
tiktoken>=0.7.0

# GraphRAG 數據科學依賴
numpy>=1.25.2
numba==0.60.0
scipy==1.12.0
networkx>=3
graspologic>=3.4.1
fastparquet>=2024.2.0

# GraphRAG 其他依賴
tenacity>=9.0.0
python-dotenv>=1.0.0
pyaml-env>=1.2.1
aiolimiter>=1.1.0
uvloop>=0.20.0
nest-asyncio>=1.6.0

# Azure 依賴（如需要）
azure-search-documents>=11.4.0
lancedb>=0.11.0
